pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "xobliam"
version = "0.1.0"
description = "Gmail analytics dashboard with intelligent cleanup recommendations"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "xobliam team"}
]
keywords = ["gmail", "email", "analytics", "cleanup"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "google-api-python-client>=2.100.0",
    "google-auth-httplib2>=0.1.0",
    "google-auth-oauthlib>=1.1.0",
    "streamlit>=1.28.0",
    "rich>=13.6.0",
    "click>=8.1.0",
    "python-dotenv>=1.0.0",
    "pandas>=2.1.0",
    "plotly>=5.18.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.10.0",
    "isort>=5.12.0",
    "flake8>=6.1.0",
    "pre-commit>=3.5.0",
]

[project.scripts]
xobliam = "xobliam.main:cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["xobliam*"]

[tool.black]
line-length = 88
target-version = ["py311"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
